Devide weather programs
